Added a the folder of the Sand Project
This has the SandLab.Java which is where all the code. the LabDisplay.class is the code for the graphics and display provided by my teacher, no way to change this code other than the size of the grid for the particles and the buttons and their labels thru the sandlab.java file.
The sandlabFileTesting is the save file and the sandlabfile is a back up for another file.
The SandLabFiles.java is the code that will save the grid and the current contents of the grid, can only save the first 9 particle types.